requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
trl>=0.7.1
agentscope>=0.1.0
tqdm>=4.65.0
pandas>=1.5.3
numpy>=1.24.3
faiss-cpu>=1.7.4
jsonlines>=3.1.0
loguru>=0.7.0
openai>=1.1.0
accelerate>=0.20.3
bitsandbytes>=0.41.0
scipy>=1.10.1
scikit-learn>=1.2.2
